requirements.txt
pyyaml
python-dotenv
pandas
matplotlib
ccxt
openai
